+++ backend/package.json
@@ -26,6 +26,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.6.1",
     "express": "^5.1.0",
+    "express-async-handler": "^1.2.0",
     "express-rate-limit": "^6.7.0",
     "helmet": "^6.1.5",
     "jsonwebtoken": "^9.0.2",
