+++ backend/routes/adminRoutes.js
@@ -9,11 +9,6 @@ const router = express.Router();
 router.use(authenticate);
 router.use(requireAdmin);
 
-/**
- * @route GET /api/admin/dashboard
- * @description Get admin dashboard overview
- * @access Private (Admin only)
- */
 router.get('/dashboard', (req, res) => {
   // In a real application, this would fetch aggregated data (e.g., user count, order count, revenue)
   res.json({
@@ -23,11 +18,6 @@ router.get('/dashboard', (req, res) => {
   });
 });
 
-/**
- * @route GET /api/admin/users
- * @description Get all users (requires 'manage_users' permission)
- * @access Private (Admin with specific permission)
- */
 router.get('/users', requirePermission('manage_users'), (req, res) => {
   // Placeholder: Implement logic to fetch all users from your database
   // Example: const users = await User.find({});
@@ -38,11 +28,6 @@ router.get('/users', requirePermission('manage_users'), (req, res) => {
   });
 });
 
-/**
- * @route POST /api/admin/users
- * @description Create a new user (requires 'manage_users' permission)
- * @access Private (Admin with specific permission)
- */
 router.post('/users', requirePermission('manage_users'), (req, res) => {
   // Placeholder: Implement logic to create a new user in your database
   // Example: const newUser = await User.create(req.body);
@@ -53,11 +38,6 @@ router.post('/users', requirePermission('manage_users'), (req, res) => {
   });
 });
 
-/**
- * @route PUT /api/admin/users/:id
- * @description Update an existing user (requires 'manage_users' permission)
- * @access Private (Admin with specific permission)
- */
 router.put('/users/:id', requirePermission('manage_users'), (req, res) => {
   // Placeholder: Implement logic to find and update a user by ID
   // Example: const updatedUser = await User.findByIdAndUpdate(req.params.id, req.body, { new: true });
@@ -68,11 +48,7 @@ router.put('/users/:id', requirePermission('manage_users'), (req, res) => {
   });
 });
 
-/**
- * @route DELETE /api/admin/users/:id
- * @description Delete a user (requires 'manage_users' permission)
- * @access Private (Admin with specific permission)
- */
+
 router.delete('/users/:id', requirePermission('manage_users'), (req, res) => {
   // Placeholder: Implement logic to find and delete a user by ID
   // Example: await User.findByIdAndDelete(req.params.id);
@@ -82,11 +58,6 @@ router.delete('/users/:id', requirePermission('manage_users'), (req, res) => {
   });
 });
 
-/**
- * @route GET /api/admin/orders
- * @description Get all orders (requires 'manage_orders' permission)
- * @access Private (Admin with specific permission)
- */
 router.get('/orders', requirePermission('manage_orders'), (req, res) => {
   // Placeholder: Implement logic to fetch all orders
   res.json({
@@ -96,11 +67,6 @@ router.get('/orders', requirePermission('manage_orders'), (req, res) => {
   });
 });
 
-/**
- * @route PUT /api/admin/orders/:id
- * @description Update order status (requires 'manage_orders' permission)
- * @access Private (Admin with specific permission)
- */
 router.put('/orders/:id', requirePermission('manage_orders'), (req, res) => {
   // Placeholder: Implement logic to update order status
   res.json({
@@ -109,11 +75,6 @@ router.put('/orders/:id', requirePermission('manage_orders'), (req, res) => {
   });
 });
 
-/**
- * @route GET /api/admin/analytics
- * @description Get analytics data (requires 'view_analytics' permission)
- * @access Private (Admin with specific permission)
- */
 router.get('/analytics', requirePermission('view_analytics'), (req, res) => {
   // Placeholder: Implement logic to fetch aggregated analytics data
   res.json({
