+++ backend/routes/authRoutes.js
@@ -1,0 +1,11 @@
+// backend/routes/authRoutes.js
+import express from 'express';
+
+import { registerUser, loginUser } from '../controllers/authController.js';
+
+const router = express.Router();
+
+router.post('/register', registerUser);
+router.post('/login', loginUser);
+
+export default router;
